refactored user query to custom hook
