+++ client/snowpack.config.mjs
@@ -5,7 +5,7 @@ export default {
     src: { url: '/dist' },
   },
   env: {
-    API_URL: 'http://localhost:4000/',
+    API_URL: 'http://localhost:3000/',
   },
   plugins: [
     '@snowpack/plugin-react-refresh',
