+++ server/server.js
@@ -19,5 +19,5 @@ app.use(errorHandler);
 
 // start server
 const port =
-  process.env.NODE_ENV === "production" ? process.env.PORT || 80 : 4000;
+  process.env.NODE_ENV === "production" ? process.env.PORT || 80 : 3000;
 app.listen(port, () => console.log("Server listening on port " + port));
